+++ packages/evm/src/wagmi-provider/index.tsx
@@ -82,7 +82,7 @@ import type {
   WalletFactory,
 } from '../interface';
 import { PelicanWeb3ConfigProvider } from './config-provider';
-import type { Token } from 'pelican-web3-lib-common';
+import type { Token, CustomToken } from 'pelican-web3-lib-common';
 import { wcConnector } from '../connector/walletConnect';
 
 /**
@@ -132,6 +132,7 @@ export interface WagmiWeb3ConfigProviderProps {
    * 指定 ERC-20 代币以查询余额（传入后优先显示该代币余额）
    */
   token?: Token;
+  customToken?: CustomToken;
   eip6963?: EIP6963Config;
   initialState?: State;
   reconnectOnMount?: boolean;
@@ -293,6 +294,7 @@ export function WagmiWeb3ConfigProvider({
           ens={ens}
           balance={balance}
           token={restProps?.token}
+          customToken={restProps?.customToken}
           eip6963={eip6963}
           wagimConfig={wagmiConfig}
           // 仅当 WalletConnect 为对象且用户开启 useWalletConnectOfficialModal 时启用官方二维码弹窗