+++ packages/tron/src/adapters/tronlink/index.ts
@@ -1,3 +1,3 @@
-export * from './adapter.js';
-export * from './types.js';
-export * from './utils.js';
+export * from './adapter';
+export * from './types';
+export * from './utils';