+++ examples/eth-web3js/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @example/eth-web3js
 
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.10
+  - pelican-web3-lib-common@1.0.10
+  - pelican-web3-lib-eth-web3js@1.0.10
+  - pelican-web3-lib@1.26.10
+
+## 0.0.35
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.9
+  - pelican-web3-lib-common@1.0.9
+  - pelican-web3-lib-eth-web3js@1.0.9
+  - pelican-web3-lib@1.26.9
+
 ## 0.0.34
 
 ### Patch Changes