+++ packages/assets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # pelican-web3-lib-assets
 
+## 1.0.10
+
+### Patch Changes
+
+- Patch release
+- Updated dependencies
+  - pelican-web3-lib-common@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- Patch release
+- Updated dependencies
+  - pelican-web3-lib-common@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes