+++ packages/ethers-v5/CHANGELOG.md
@@ -1,5 +1,27 @@
 # pelican-web3-lib-ethers-v5
 
+## 1.0.10
+
+### Patch Changes
+
+- Patch release
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.10
+  - pelican-web3-lib-common@1.0.10
+  - pelican-web3-lib-ethers@1.0.10
+  - pelican-web3-lib-evm@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- Patch release
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.9
+  - pelican-web3-lib-common@1.0.9
+  - pelican-web3-lib-ethers@1.0.9
+  - pelican-web3-lib-evm@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes