+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,21 @@
 # pelican-web3-lib
 
+## 1.26.10
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.10
+  - pelican-web3-lib-common@1.0.10
+
+## 1.26.9
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.9
+  - pelican-web3-lib-common@1.0.9
+
 ## 1.26.8
 
 ### Patch Changes